dev-requirements.txt
-e .[async,test]
doc8
flake8
flake8-future-import
pep8-naming
plumbum
pyenchant
pytest-cov
shovel
sphinx
sphinx_rtd_theme
sphinxcontrib-spelling
tox
twine
watchdog